Properly track multiple snapshots per test file, add more tests, remove babel as dependency as it's only required for inline snapshots
